Move LayerManager to core package
