+++ portal/core/layer_manager.py
